requirements.txt
pandas
numpy
SQLAlchemy
python-dotenv
requests
statsmodels
scikit-learn
rapidfuzz
matplotlib
streamlit
